Split utils into different files for clarity and updated imports
