+++ src/utils/pagination.ts
@@ -1,0 +1,36 @@
+export function generatePagination(currentPage: number, totalPages: number) {
+  // Aims to return 7 slots if possible
+  // Return all pages if 7 or fewer
+  if (totalPages <= 7) {
+    return Array.from({ length: totalPages }, (_, i) => i + 1);
+  }
+
+  // If current page in the first three, show ellipsis after 4
+  if (currentPage <= 3) {
+    return [1, 2, 3, 4, '...', totalPages - 1, totalPages];
+  }
+
+  // If current page in the last four, show ellipsis before last 4
+  if (currentPage >= totalPages - 2) {
+    return [
+      1,
+      2,
+      '...',
+      totalPages - 3,
+      totalPages - 2,
+      totalPages - 1,
+      totalPages,
+    ];
+  }
+
+  // If current page in the middle, show neighbours, first and last
+  return [
+    1,
+    '...',
+    currentPage - 1,
+    currentPage,
+    currentPage + 1,
+    '...',
+    totalPages,
+  ];
+}
